Move `SocketEvent` and `SocketEventInit` to the Switch namespace
